feat: add `DataInterpolationsOptimExt`
- Whenever Optim is found in the env, Curvefit and corr. _interpolate dispatches are loaded
