+++ src/interpolation_caches.jl
@@ -354,35 +354,3 @@ function BSplineApprox(u,t,d,h,pVecType,knotVecType)
   c[2:end-1] .= vec(P)
   BSplineApprox{true}(u,t,d,h,p,k,c,pVecType,knotVecType)
 end
-
-
-### Curvefit
-struct CurvefitCache{uType,tType,mType,p0Type,ubType,lbType,algType,pminType,FT,T} <: AbstractInterpolation{FT,T}
-  u::uType
-  t::tType
-  m::mType        # model type
-  p0::p0Type      # intial params
-  ub::ubType      # upper bound of params
-  lb::lbType      # lower bound of params
-  alg::algType    # alg to optimize cost function
-  pmin::pminType  # optimized params
-  CurvefitCache{FT}(u,t,m,p0,ub,lb,alg,pmin) where FT = new{typeof(u),typeof(t),typeof(m),
-                                      typeof(p0),typeof(ub),typeof(lb),
-                                      typeof(alg),typeof(pmin),FT,eltype(u)}(u,t,m,p0,ub,lb,alg,pmin)
-end
-
-function Curvefit(u, t, model, p0, alg, box=false, lb=nothing, ub=nothing)
-  u, t = munge_data(u, t)
-  errfun(t,u,p) = sum(abs2.(u .- model(t,p)))
-  if box == false
-      mfit = optimize(p->errfun(t,u,p), p0, alg)
-  else
-      if lb == nothing || ub == nothing
-          error("lower or upper bound should not be nothing")
-      end
-      od = OnceDifferentiable(p->errfun(t,u,p), p0, autodiff=:finite)
-      mfit = optimize(od, lb, ub, p0, Fminbox(alg))
-  end
-  pmin = Optim.minimizer(mfit)
-  CurvefitCache{true}(u,t,model,p0,ub,lb,alg,pmin)
-end
